requirements.txt
pyyaml==5.4.1
jericho
spacy==3.1.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0.tar.gz
langid==1.1.6
emoji==1.4.2
